updated the models, organized file structure
