+++ server.js
@@ -6,6 +6,8 @@ const routes = require("./controllers");
 
 const PORT = process.env.PORT || 3000;
 
+
+
 const app = express();
 
 app.use(logger("dev"));
@@ -13,21 +15,29 @@ app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 app.use(express.static("public"));
 
+// local
+
+const cors = require("cors")
+
+//DEPLOYED
+// heroku created nothing pushed yet
+
+// app.use(cors({
+//   origin:["https://there-and-back-again-front.herokuapp.com/"]
+// }));
+
+app.use('/',allRoutes);
+
 
 // routes
-app.use(require("./controllers/api"))
+app.use(require("./controllers/index2"))
 app.use(require("./controllers/"))
 
 app.use("/", routes);
 
 
 
-// mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/trail", {
-//   useNewUrlParser: true,
-//   useUnifiedTopology: true,
-//   useCreateIndex: true,
-//   useFindAndModify: false,
-// });
+
 
 app.listen(PORT, () => {
   console.log(`App running on port ${PORT}!`);
